Add files via upload
A2DGaugeReader.py is stored in the spot_cam directory. Once the script is started is listens for new .jpg files in the spot_cam directory. This directory is where new .jpg images from Spot are placed. Analysis follows. MQtt messages are sent to the message broker and these are used to generate work orders etc. in Maximo